requirements.txt
bcrypt
fastapi
uvicorn
twilio
python-dotenv
python-multipart
google-auth
google-auth-oauthlib
google-auth-httplib2
gspread
openai
pandas
thefuzz
rapidfuzz
requests
python-magic
PyPDF2
selenium
